backend/requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
python-decouple==3.8

# Database
psycopg2-binary==2.9.9

# CORS handling
django-cors-headers==4.3.1

# API enhancements
django-filter==23.5
drf-spectacular==0.27.0

# Caching & Tasks
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
django-celery-beat==2.5.0

# Payments
stripe==7.10.0

# Emails
requests==2.31.0

# PDF Generation
weasyprint==60.2

# Development
django-debug-toolbar==4.2.0

# Production server
gunicorn==21.2.0
whitenoise==6.6.0
